LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-media-dynamic-expected.txt



FAIL Custom Functions: @media responds to changes assert_equals: --actual after resize to 100px expected "B" but got "A"
